+++ app/containers/wallet/dialogs/WalletKeyImportDialogContainer.js
@@ -7,11 +7,11 @@ import type { InjectedDialogContainerProps } from '../../../types/injectedPropsT
 @inject('stores', 'actions') @observer
 export default class WalletKeyImportDialogContainer extends Component {
 
-  static defaultProps = { actions: null, stores: null, children: null };
+  static defaultProps = { actions: null, stores: null, children: null, onClose: () => {} };
 
   props: InjectedDialogContainerProps;
 
-  onSubmit = (values: { filePath: string }) => {
+  onSubmit = (values: { filePath: string, walletPassword: ?string }) => {
     this.props.actions.wallets.importWalletFromKey.trigger(values);
   };